requirements.txt
torch
transformers
pandas
scikit-learn
datasets
accelerate
fastapi
uvicorn
torch
transformers
pydantic
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-jose
passlib[bcrypt]
alembic
requests